School-PHP1_2
The project is the realization of a one-page site that was implemented using procedural PHP and MySQL database. The menu is dynamically printed from the database depending on the role of the logged user. The survey and the administration of the survey were realized with the help of PHP and Ajax. The gallery was realized with the help of the JQuery library valillabox, while a page Prices was realized with pagination. Login can be realized through two roles, user and administrator. Login, filling in the scheduling form, as well as adding and changing during the administration of the content by the administrator was realized by checking with regular expressions on the server side and with the help of procedural PHP. The user role accesses to the pages scheduled, canceled and contact. The user is able to send an email to the administrator. The Administrator role has access to the admin panel and is able to do administration through action such as adding, modifying, and deleting content on pages. The administrator is able to do administration of content on pages such as prices, appointments and cancellations, surveys, services and users.
Technologies – HTML5, CSS3, PHP, JavaScript, JQuery – vanillabox, Ajax, PHPMailer